adds be for update of internship

+++ backend/backend/backend/src/Application/InternShips/Commands/CreateInternShip/InternShipCreateDto.cs
@@ -11,7 +11,7 @@ public class InternShipCreateDto
 {
     public string SchoolYear { get; set; }
     public int UnitId { get; set; }
-    public byte MaxCountOfStudents { get; set; }
+    public int MaxCountOfStudents { get; set; }
     public int CurrentCountOfStudents { get; set; }
     public TrainingType TrainingType { get; set; }
